test(core): update test cases
